feat: wip added dynamic filters
