+++ src/components/Product/hooks/useProductFilters.ts
@@ -1,76 +1,14 @@
-import { useMemo, useState } from 'react';
-import { BatchFilterItems } from '../helpers';
+import { useMemo } from "react";
 
 type UseProductFiltersParams = {
-  institutions: Array<any>;
-  batchFilterItems: Array<BatchFilterItems>;
+  setFilters: (filter: Record<string, {value: string; label: string}>) => void;
+  filters: Record<string, {value: string; label: string}>;
 };
 
-export const useProductFilters = ({ institutions, batchFilterItems }: UseProductFiltersParams) => {
-  const [categoryFilter, setCategoryFilter] = useState<string>('');
-  const [producerFilter, setProducerFilter] = useState<string>('');
-  const [batchFilter, setBatchFilter] = useState<string>('');
-  const [statusFilter, setStatusFilter] = useState<string>('');
-  const [eprelCodeFilter, setEprelCodeFilter] = useState<string>('');
-  const [gtinCodeFilter, setGtinCodeFilter] = useState<string>('');
+export const useProductFilters = ({ setFilters, filters }: UseProductFiltersParams) => {
+  const filtersLabel = useMemo(() => Object.values(filters).reduce((acc, value) => `${acc ? ',' : ''} ${value?.label}`, ''), [filters]);
 
-  const resetFilters = () => {
-    setCategoryFilter('');
-    setStatusFilter('');
-    setProducerFilter('');
-    setBatchFilter('');
-    setEprelCodeFilter('');
-    setGtinCodeFilter('');
-  };
+  const resetFilters = () => setFilters({});
 
-  const filtersLabel = useMemo(() => {
-    const norm = (s?: string) => (s ? s.trim() : '');
-
-    const producer = producerFilter?.trim()
-      ? institutions?.find((p: any) => p.institutionId === producerFilter)?.description ||
-        producerFilter.trim()
-      : '';
-
-    const batch = batchFilter?.trim()
-      ? batchFilterItems?.find((b) => b?.productFileId === batchFilter)?.batchName ||
-        batchFilter.trim()
-      : '';
-
-    return [
-      norm(categoryFilter),
-      norm(statusFilter),
-      norm(producer),
-      norm(batch),
-      norm(eprelCodeFilter),
-      norm(gtinCodeFilter),
-    ]
-      .filter((s): s is string => !!s)
-      .join(', ');
-  }, [
-    categoryFilter,
-    statusFilter,
-    producerFilter,
-    institutions,
-    batchFilter,
-    batchFilterItems,
-    eprelCodeFilter,
-    gtinCodeFilter,
-  ]);
-
-  return {
-    categoryFilter,
-    setCategoryFilter,
-    producerFilter,
-    setProducerFilter,
-    batchFilter,
-    setBatchFilter,
-    statusFilter,
-    setStatusFilter,
-    eprelCodeFilter,
-    setEprelCodeFilter,
-    gtinCodeFilter,
-    setGtinCodeFilter,
-    filtersLabel,
-    resetFilters,
-  };
+  return { filtersLabel, resetFilters };
 };
