+++ src/locale/it/bonusDecoder2026/copy.json
@@ -25,22 +25,44 @@
         "productCode": "Codice prodotto",
         "category": "Categoria",
         "status": "Stato",
-        "batchName": "Lotto"
+        "productFileId": "Lotto"
       },
       "filters": {
         "category": "Categoria",
         "status": "Stato",
         "gtinCode": "Codice GTIN/EAN",
         "productCode": "Codice prodotto",
-        "productFileId": "Lotto"
+        "productFileId": "Lotto",
+        "errors": {
+          "gtinCode": "Il codice deve avere 14 caratteri"
+        }
       }
     }
   },
-  "categories": {
-    "satellite": { "label": "Satellitare" },
-    "terrestrial": { "label": "Terrestre" },
-    "terrestrial_via_cable": { "label": "Terrestre via Cavo" },
-    "terrestrial_and_satellite": { "label": "Terrestre e Satellitare" },
-    "terrestrial_satellite_and_via_cable": { "label": "Terrestre, Satellitare e via Cavo" }
+  "category": {
+    "satellite": {
+      "label": "Satellitare"
+    },
+    "terrestrial": {
+      "label": "Terrestre"
+    },
+    "terrestrial_via_cable": {
+      "label": "Terrestre via Cavo"
+    },
+    "terrestrial_and_satellite": {
+      "label": "Terrestre e Satellitare"
+    },
+    "terrestrial_satellite_and_via_cable": {
+      "label": "Terrestre, Satellitare e via Cavo"
+    }
+  },
+  "status": {
+    "products": {
+      "UPLOADED": "Da Esaminare",
+      "WAIT_APPROVED": "Da approvare",
+      "SUPERVISED": "Da controllare",
+      "APPROVED": "Approvato",
+      "REJECTED": "Escluso"
+    }
   }
 }
