+++ src/locale/it/bonusElettrodomestici2025/default/config.json
@@ -6,13 +6,17 @@
       "default": {
         "logicalName": "Utente Standard",
         "permissions": {
-          "tables": ["products"]
+          "tables": [
+            "products"
+          ]
         }
       },
       "super_admin": {
         "logicalName": "Super Admin",
         "permissions": {
-          "tables": ["products"]
+          "tables": [
+            "products"
+          ]
         }
       }
     },
@@ -25,22 +29,57 @@
     "tables": {
       "products": {
         "columns": [
-          { "id": "category", "labelKey": "tables.products.columns.category", "sortable": true },
+          {
+            "id": "category",
+            "labelKey": "tables.products.columns.category",
+            "sortable": true
+          },
           {
             "id": "energyClass",
             "labelKey": "tables.products.columns.energyClass",
             "sortable": true
           },
-          { "id": "eprelCode", "labelKey": "tables.products.columns.eprelCode", "sortable": true },
-          { "id": "gtinCode", "labelKey": "tables.products.columns.gtinCode", "sortable": true },
-          { "id": "batchName", "labelKey": "tables.products.columns.batchName", "sortable": false },
-          { "id": "status", "labelKey": "tables.products.columns.status", "sortable": true }
+          {
+            "id": "eprelCode",
+            "labelKey": "tables.products.columns.eprelCode",
+            "sortable": true
+          },
+          {
+            "id": "gtinCode",
+            "labelKey": "tables.products.columns.gtinCode",
+            "sortable": true
+          },
+          {
+            "id": "productFileId",
+            "labelKey": "tables.products.columns.productFileId",
+            "sortable": false
+          },
+          {
+            "id": "status",
+            "labelKey": "tables.products.columns.status",
+            "sortable": true
+          }
         ],
-        "defaultSort": { "field": "gtinCode", "direction": "asc" },
+        "defaultSort": {
+          "field": "gtinCode",
+          "direction": "asc"
+        },
         "filters": [
-          { "id": "category", "type": "select", "labelKey": "tables.products.filters.category" },
-          { "id": "status", "type": "select", "labelKey": "tables.products.filters.status" },
-          { "id": "gtinCode", "type": "text", "labelKey": "tables.products.filters.gtinCode" },
+          {
+            "id": "category",
+            "type": "select",
+            "labelKey": "tables.products.filters.category"
+          },
+          {
+            "id": "status",
+            "type": "select",
+            "labelKey": "tables.products.filters.status"
+          },
+          {
+            "id": "gtinCode",
+            "type": "text",
+            "labelKey": "tables.products.filters.gtinCode"
+          },
           {
             "id": "productCode",
             "type": "text",
@@ -53,20 +92,41 @@
           }
         ],
         "ui": {
-          "filtering": { "enabled": true },
+          "filtering": {
+            "enabled": true
+          },
           "pagination": {
             "enabled": true,
             "defaultRowsPerPage": 10,
-            "rowsPerPageOptions": [10, 25, 50]
+            "rowsPerPageOptions": [
+              10,
+              25,
+              50
+            ]
           }
         },
         "detail": {
           "fields": [
-            { "id": "gtinCode", "labelKey": "pages.productDetail.gtinCode" },
-            { "id": "productCode", "labelKey": "pages.productDetail.productCode" },
-            { "id": "brand", "labelKey": "pages.productDetail.brand" },
-            { "id": "model", "labelKey": "pages.productDetail.model" },
-            { "id": "status", "labelKey": "pages.productDetail.status" }
+            {
+              "id": "gtinCode",
+              "labelKey": "pages.productDetail.gtinCode"
+            },
+            {
+              "id": "productCode",
+              "labelKey": "pages.productDetail.productCode"
+            },
+            {
+              "id": "brand",
+              "labelKey": "pages.productDetail.brand"
+            },
+            {
+              "id": "model",
+              "labelKey": "pages.productDetail.model"
+            },
+            {
+              "id": "status",
+              "labelKey": "pages.productDetail.status"
+            }
           ]
         }
       }
