+++ src/locale/it/bonusElettrodomestici2025/operatore/config.json
@@ -6,7 +6,9 @@
       "operatore": {
         "logicalName": "Produttore",
         "permissions": {
-          "tables": ["products"]
+          "tables": [
+            "products"
+          ]
         }
       }
     }
@@ -14,23 +16,93 @@
   "templates": {
     "formats": {
       "csv": {
-        "headers": ["Codice GTIN/EAN", "Codice Prodotto", "Categoria", "Marca", "Modello"],
-        "fields": ["GTIN123ABC4567", "SKU98765", "{{category}}", "MarcaEsempio", "ModelloX100"]
+        "headers": [
+          "Codice GTIN/EAN",
+          "Codice Prodotto",
+          "Categoria",
+          "Marca",
+          "Modello"
+        ],
+        "fields": [
+          "GTIN123ABC4567",
+          "SKU98765",
+          "{{category}}",
+          "MarcaEsempio",
+          "ModelloX100"
+        ]
       },
       "eprel": {
-        "headers": ["Codice EPREL","Codice GTIN/EAN", "Codice Prodotto", "Categoria", "Paese di Produzione"],
-        "fields": ["12345", "GTIN01234", "SKU12345", "Lavatrice", "IT"]
+        "headers": [
+          "Codice EPREL",
+          "Codice GTIN/EAN",
+          "Codice Prodotto",
+          "Categoria",
+          "Paese di Produzione"
+        ],
+        "fields": [
+          "12345",
+          "GTIN01234",
+          "SKU12345",
+          "Lavatrice",
+          "IT"
+        ]
       }
     },
-    "categories": {
-      "washingmachines": { "format": "eprel", "name": "eprel" },
-      "tumbledryers": { "format": "eprel", "name": "eprel" },
-      "ovens": { "format": "eprel", "name": "eprel" },
-      "rangehoods": { "format": "eprel", "name": "eprel" },
-      "dishwashers": { "format": "eprel", "name": "eprel" },
-      "washerdriers": { "format": "eprel", "name": "eprel" },
-      "refrigeratingappl": { "format": "eprel", "name": "eprel" },
-      "cookinghobs": { "format": "csv", "name": "cookinghobs" }
+    "category": {
+      "washingmachines": {
+        "label": "category.washingmachines.label",
+        "format": "eprel",
+        "name": "eprel"
+      },
+      "tumbledryers": {
+        "label": "category.tumbledryers.label",
+        "format": "eprel",
+        "name": "eprel"
+      },
+      "ovens": {
+        "label": "category.ovens.label",
+        "format": "eprel",
+        "name": "eprel"
+      },
+      "rangehoods": {
+        "label": "category.rangehoods.label",
+        "format": "eprel",
+        "name": "eprel"
+      },
+      "dishwashers": {
+        "label": "category.dishwashers.label",
+        "format": "eprel",
+        "name": "eprel"
+      },
+      "washerdriers": {
+        "label": "category.washerdriers.label",
+        "format": "eprel",
+        "name": "eprel"
+      },
+      "refrigeratingappl": {
+        "label": "category.refrigeratingappl.label",
+        "format": "eprel",
+        "name": "eprel"
+      },
+      "cookinghobs": {
+        "label": "category.cookinghobs.label",
+        "format": "csv",
+        "name": "cookinghobs"
+      }
+    },
+    "status": {
+      "UPLOADED": {
+        "label": "status.products.UPLOADED",
+        "color": "default"
+      },
+      "APPROVED": {
+        "label": "status.products.APPROVED",
+        "color": "success"
+      },
+      "REJECTED": {
+        "label": "status.products.REJECTED",
+        "color": "error"
+      }
     }
   },
   "ui": {
@@ -63,8 +135,8 @@
             "align": "center"
           },
           {
-            "id": "batchName",
-            "labelKey": "tables.products.columns.batchName",
+            "id": "productFileId",
+            "labelKey": "tables.products.columns.productFileId",
             "sortable": false,
             "align": "left"
           },
@@ -75,13 +147,53 @@
             "align": "left"
           }
         ],
-        "defaultSort": { "field": "gtinCode", "direction": "asc" },
+        "defaultSort": {
+          "field": "gtinCode",
+          "direction": "asc"
+        },
+        "filters": [
+          {
+            "id": "status",
+            "type": "select",
+            "labelKey": "tables.products.filters.status"
+          },
+          {
+            "id": "productFileId",
+            "type": "select",
+            "labelKey": "tables.products.filters.productFileId"
+          },
+          {
+            "id": "category",
+            "type": "select",
+            "labelKey": "tables.products.filters.category"
+          },
+          {
+            "id": "eprelCode",
+            "type": "text",
+            "labelKey": "tables.products.filters.eprelCode",
+            "regEx": "^\/d+$",
+            "error": "tables.products.filters.errors.eprelCode"
+          },
+          {
+            "id": "gtinCode",
+            "type": "text",
+            "labelKey": "tables.products.filters.gtinCode",
+            "regEx": "^[a-zA-Z0-9]{1,14}$",
+            "error": "tables.products.filters.errors.gtinCode"
+          }
+        ],
         "ui": {
-          "filtering": { "enabled": true },
+          "filtering": {
+            "enabled": true
+          },
           "pagination": {
             "enabled": true,
             "defaultRowsPerPage": 10,
-            "rowsPerPageOptions": [10, 25, 50]
+            "rowsPerPageOptions": [
+              10,
+              25,
+              50
+            ]
           }
         }
       }
